# packages/vibeship-mind/vibeship_mind-5.0.1.tar.gz/vibeship_mind-5.0.1/docs/specs/openapi.yaml
openapi: 3.1.0
info:
  title: Mind v5 API
  description: |
    Decision intelligence API for AI agents. Mind v5 provides hierarchical temporal memory,
    outcome-weighted retrieval, causal reasoning, and collective intelligence through
    federated pattern learning.

    ## Core Concepts

    - **Memories**: Hierarchical temporal memories with outcome-weighted salience
    - **Decisions**: Tracked decision traces with context snapshots
    - **Outcomes**: Feedback signals that adjust memory salience for future retrieval
    - **Context**: Fused retrieval from vector, graph, and keyword sources

    ## Authentication

    All endpoints require Bearer JWT authentication. Tokens are scoped to a single user
    and include the user_id claim.

    ## Rate Limits

    - Standard: 1000 requests/minute
    - Burst: 10000 requests/minute (short bursts)
    - Memory retrieval: 100 requests/second per user

    ## Versioning

    This API follows semantic versioning. Breaking changes require a major version bump.
    The current version is v1. Use the `Accept-Version` header for version negotiation.
  version: 1.0.0
  contact:
    name: Mind v5 Team
    email: mind-api@example.com
    url: https://docs.mind-v5.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.mind-v5.dev/v1
    description: Production server
  - url: https://staging-api.mind-v5.dev/v1
    description: Staging server
...
security:
  - BearerAuth: []
...
paths:
...
  /decisions/track:
    post:
      operationId: trackDecision
      summary: Track a decision with memories used
      description: |
        Records a decision trace including:
        - Context snapshot (memories that were retrieved and used)
        - Decision made (type, content, confidence)
        - Alternatives considered

        Returns a trace_id that should be used when observing outcomes.
        This enables the system to attribute outcomes to specific memories
        and adjust their salience accordingly.

        **Flow**:
        1. Retrieve context via `/memories/retrieve`
        2. Make decision using retrieved memories
        3. Track decision via this endpoint
        4. Later, observe outcome via `/outcomes/observe`
      tags:
        - Decisions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionTrackRequest'
            examples:
              recommendation:
                summary: Recommendation decision
                value:
                  session_id: "550e8400-e29b-41d4-a716-446655440000"
                  decision_type: "recommendation"
                  decision_content: "Suggested using Temporal.io for workflow orchestration"
                  confidence: 0.87
                  context_memory_ids:
                    - "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                    - "b2c3d4e5-f6a7-8901-bcde-f12345678901"
                  memory_weights:
                    a1b2c3d4-e5f6-7890-abcd-ef1234567890: 0.8
                    b2c3d4e5-f6a7-8901-bcde-f12345678901: 0.6
                  alternatives_considered:
                    - "Celery with Redis"
                    - "Apache Airflow"
                  metadata:
                    domain: "architecture"
                    complexity: "high"
      responses:
        '201':
          description: Decision tracked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionTrace'
              examples:
                tracked:
                  summary: Tracked decision
                  value:
                    trace_id: "d4e5f6a7-b8c9-0123-def0-456789abcdef"
                    session_id: "550e8400-e29b-41d4-a716-446655440000"
                    decision_type: "recommendation"
                    decision_content: "Suggested using Temporal.io for workflow orchestration"
                    confidence: 0.87
                    context_memory_ids:
                      - "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                      - "b2c3d4e5-f6a7-8901-bcde-f12345678901"
                    outcome_observed: false
                    created_at: "2025-12-27T19:45:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /decisions/{trace_id}:
    get:
      operationId: getDecisionTrace
      summary: Get a decision trace
      description: |
        Retrieves a decision trace by its ID. Includes the full context
        snapshot, decision details, and outcome if observed.
      tags:
        - Decisions
      parameters:
        - name: trace_id
          in: path
          required: true
          description: Unique identifier of the decision trace
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Decision trace found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionTrace'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /context:
    get:
      operationId: getContextSummary
      summary: Get current user context summary
      description: |
        Returns a summary of the user's current context across all temporal
        levels. Useful for understanding the user's state at a glance.

        Includes:
        - Active memories by temporal level
        - Recent decision success rate (DSR)
        - Top entities and relationships
        - Pending decisions awaiting outcome
      tags:
        - Context
      parameters:
        - name: session_id
          in: query
          required: false
          description: Filter to specific session context
          schema:
            type: string
            format: uuid
        - name: include_metrics
          in: query
          required: false
          description: Include performance metrics
          schema:
            type: boolean
            default: true
        - name: include_pending_decisions
          in: query
          required: false
          description: Include decisions awaiting outcomes
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Context summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextSummary'
              examples:
                full:
                  summary: Full context summary
                  value:
                    user_id: "123e4567-e89b-12d3-a456-426614174000"
                    snapshot_timestamp: "2025-12-27T20:15:00Z"
                    memory_counts:
                      immediate: 12
                      situational: 45
                      seasonal: 23
                      identity: 8
                    total_memories: 88
                    active_session_id: "550e8400-e29b-41d4-a716-446655440000"
                    metrics:
                      decision_success_rate: 0.78
                      total_decisions: 156
                      positive_outcomes: 122
                      negative_outcomes: 34
                      avg_retrieval_latency_ms: 42.5
                    top_entities:
                      - name: "Mind v5"
                        type: "project"
                        mention_count: 47
                      - name: "Temporal.io"
                        type: "technology"
                        mention_count: 23
                    pending_decisions_count: 3
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token with the following claims:
        - `sub`: User ID (UUID)
        - `iat`: Issued at timestamp
        - `exp`: Expiration timestamp
        - `scope`: Allowed operations (read, write, admin)
  schemas:
...
    DecisionTrackRequest:
      type: object
      description: Request to track a decision
      required:
        - decision_type
        - decision_content
        - confidence
        - context_memory_ids
      properties:
        session_id:
          type: string
          format: uuid
          description: Session in which decision was made
        decision_type:
          type: string
          description: Category of decision
          maxLength: 100
          example: recommendation
        decision_content:
          type: string
          description: The decision that was made
          maxLength: 10000
          example: "Suggested using Temporal.io for workflow orchestration"
        confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: Confidence in the decision
          example: 0.87
        context_memory_ids:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1
          description: Memory IDs that were used for this decision
          example:
            - "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
            - "b2c3d4e5-f6a7-8901-bcde-f12345678901"
        memory_weights:
          type: object
          additionalProperties:
            type: number
            format: float
          description: How much each memory contributed (memory_id -> weight)
          example:
            a1b2c3d4-e5f6-7890-abcd-ef1234567890: 0.8
            b2c3d4e5-f6a7-8901-bcde-f12345678901: 0.6
        alternatives_considered:
          type: array
          items:
            type: string
          description: Other options that were considered
          example:
            - "Celery with Redis"
            - "Apache Airflow"
        metadata:
          type: object
          additionalProperties: true
          description: Additional decision context
    DecisionTrace:
      type: object
      description: A tracked decision with context snapshot
      required:
        - trace_id
        - decision_type
        - decision_content
        - confidence
        - context_memory_ids
        - outcome_observed
        - created_at
      properties:
        trace_id:
          type: string
          format: uuid
          description: Unique trace identifier
          example: "d4e5f6a7-b8c9-0123-def0-456789abcdef"
        session_id:
          type: string
          format: uuid
          nullable: true
        decision_type:
          type: string
          example: recommendation
        decision_content:
          type: string
          example: "Suggested using Temporal.io for workflow orchestration"
        confidence:
          type: number
          format: float
          example: 0.87
        context_memory_ids:
          type: array
          items:
            type: string
            format: uuid
        context_snapshot:
          type: object
          description: Full snapshot of context at decision time
          additionalProperties: true
        memory_weights:
          type: object
          additionalProperties:
            type: number
            format: float
        alternatives_considered:
          type: array
          items:
            type: string
        outcome_observed:
          type: boolean
          description: Whether outcome has been recorded
          example: false
        outcome_quality:
          type: number
          format: float
          nullable: true
          description: Outcome quality if observed
        outcome_timestamp:
          type: string
          format: date-time
          nullable: true
        outcome_signal:
          type: string
          nullable: true
          description: How outcome was detected
        memory_attribution:
          type: object
          additionalProperties:
            type: number
            format: float
          nullable: true
          description: Computed attribution scores
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
          example: "2025-12-27T19:45:00Z"

    # === Outcome Schemas ===
...
    ContextSummary:
      type: object
      description: Summary of user's current context
      required:
        - user_id
        - snapshot_timestamp
        - memory_counts
        - total_memories
      properties:
        user_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        snapshot_timestamp:
          type: string
          format: date-time
          example: "2025-12-27T20:15:00Z"
        memory_counts:
          type: object
          description: Memory count by temporal level
          properties:
            immediate:
              type: integer
              example: 12
            situational:
              type: integer
              example: 45
            seasonal:
              type: integer
              example: 23
            identity:
              type: integer
              example: 8
          required:
            - immediate
            - situational
            - seasonal
            - identity
        total_memories:
          type: integer
          description: Total active memories
          example: 88
        active_session_id:
          type: string
          format: uuid
          nullable: true
          description: Current active session
        metrics:
          $ref: '#/components/schemas/UserMetrics'
        top_entities:
          type: array
          items:
            $ref: '#/components/schemas/EntitySummary'
          description: Most mentioned entities
        pending_decisions_count:
          type: integer
          description: Decisions awaiting outcomes
          example: 3
        pending_decisions:
          type: array
          items:
            $ref: '#/components/schemas/DecisionTrace'
          description: Full pending decision details (if requested)
    UserMetrics:
      type: object
      description: User performance metrics
      properties:
        decision_success_rate:
          type: number
          format: float
          description: DSR - decisions with positive outcomes / total
          example: 0.78
        total_decisions:
          type: integer
          example: 156
        positive_outcomes:
          type: integer
          example: 122
        negative_outcomes:
          type: integer
          example: 34
        avg_retrieval_latency_ms:
          type: number
          format: float
          example: 42.5
        memory_effectiveness:
          type: object
          additionalProperties:
            type: number
            format: float
          description: Effectiveness by memory type
    EntitySummary:
      type: object
      description: Summary of a mentioned entity
      properties:
        name:
          type: string
          example: "Mind v5"
        type:
          type: string
          example: "project"
        mention_count:
          type: integer
          example: 47

    # === Error Schemas ===
    Error:
      type: object
      description: Standard error response
      required:
        - code
        - message
        - request_id
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: "MEMORY_NOT_FOUND"
        message:
          type: string
          description: Human-readable error message
          example: "Memory a1b2c3d4-e5f6-7890-abcd-ef1234567890 not found"
        request_id:
          type: string
          description: Request ID for debugging
          example: "req_abc123"
        details:
          type: object
          additionalProperties: true
          description: Additional error context
    ValidationError:
      type: object
      description: Validation error with field details
      required:
        - code
        - message
        - request_id
        - errors
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Request validation failed"
        request_id:
          type: string
          example: "req_abc123"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "query"
              message:
                type: string
                example: "Query must not be empty"
              code:
                type: string
                example: "required"
  responses:
    BadRequest:
      description: Bad request - malformed input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "BAD_REQUEST"
            message: "Request body is not valid JSON"
            request_id: "req_abc123"
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "UNAUTHORIZED"
            message: "Invalid or expired authentication token"
            request_id: "req_abc123"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "NOT_FOUND"
            message: "The requested resource was not found"
            request_id: "req_abc123"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            code: "VALIDATION_ERROR"
            message: "Request validation failed"
            request_id: "req_abc123"
            errors:
              - field: "query"
                message: "Query must not be empty"
                code: "required"
    RateLimited:
      description: Too many requests
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per minute
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when limit resets
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "RATE_LIMITED"
            message: "Too many requests. Please retry after 30 seconds."
            request_id: "req_abc123"
            details:
              retry_after: 30
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "INTERNAL_ERROR"
            message: "An unexpected error occurred. Please try again."
            request_id: "req_abc123"